refactor: move pyarrow out of common utils and address additional review commetns
